% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/awscli2-package.R
\docType{package}
\name{awscli2-package}
\alias{awscli2}
\alias{awscli2-package}
\title{awscli2: Wraps the AWS CLI}
\description{
Wraps the AWS CLI (v2) in a lightweight and consistent manner.
}
\author{
\strong{Maintainer}: Murat Tasan \email{mmuurr@gmail.com} (\href{https://orcid.org/0000-0002-8155-2565}{ORCID})

}
\keyword{internal}
